# openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Voice SDR API
  version: 1.0.0
  description: |
    Esta API permite configurar, testar e operar um agente de vendas por voz baseado em inteligência artificial. O agente interpreta o que o lead disse, decide o próximo passo comercial e gera uma resposta falada. A API permite ajustar e testar o comportamento antes de realizar chamadas reais.

    Com esta API você pode configurar o comportamento e as instruções do agente, escolher e ajustar a voz, configurar decisões comerciais, testar frases sem telefonar, ouvir o áudio gerado, medir o tempo de resposta, consultar o estado dos serviços, iniciar e acompanhar chamadas e consultar transcrições.

    Fluxo recomendado: (1) faça login; (2) consulte GET /v1/runtime/status para verificar o estado do sistema; (3) confira ou altere o prompt; (4) ajuste decisão comercial, voz e estilo de fala; (5) teste uma frase de lead; (6) confira a decisão e execute um teste de resposta do agente; (7) ouça o áudio; (8) ajuste configurações se necessário; (9) somente depois realize uma chamada real. Os testes sem chamada permitem avaliar intenção, próximo passo, latência, resposta gerada e áudio.

    O mecanismo de decisão comercial seleciona um próximo passo, como continuar, perguntar, tratar uma objeção, propor uma ação, fazer follow-up, encerrar ou solicitar atendimento humano. Em seguida, o modelo de voz gera a resposta. Isso separa estratégia comercial de personalidade, estilo, voz, ritmo e tom.

...

...

...
servers:
...
paths:
...
  /v1/config/sip-trunk:
    get:
      operationId: getFalePacoSIPProfile
      tags: [Fale Paco telephony]
      summary: Read the effective Fale Paco Baresip profile
      description: Returns the server-controlled canonical profile and whether a password is configured. The password and SIP authentication material are never returned.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Effective safe Fale Paco profile, content: { application/json: { schema: { $ref: '#/components/schemas/FalePacoSIPProfile' }, example: { provider: 'Fale Paco', host: '98034.falepaco.com.br', port: 5060, transport: tcp, registrar: '98034.falepaco.com.br', outbound_proxy: '98034.falepaco.com.br:5060', auth: { type: userpass, username: '100', realm: '98034.falepaco.com.br' }, from_user: '100', from_domain: '98034.falepaco.com.br', caller_id: '551155200455', registration_required: true, enabled: true, password_configured: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: Baresip profile unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    put:
      operationId: setFalePacoSIPPassword
      tags: [Fale Paco telephony]
      summary: Set the Fale Paco SIP password
      description: All provider, host, port, transport, username, realm, registrar, proxy, caller ID, and registration fields are server-controlled. Store the supplied password only in the protected local Baresip runtime account file. This operation restarts Baresip to apply the profile and sends no call.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        description: The owner's SIP password. No other SIP fields are accepted.
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FalePacoSIPPassword' }
            example: { secret: YOUR_SIP_PASSWORD }
      responses:
        '200': { description: Safe effective profile; secret is omitted, content: { application/json: { schema: { $ref: '#/components/schemas/FalePacoSIPProfile' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { description: An active call prevents restarting Baresip, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { $ref: '#/components/responses/RuntimeFailure' }
  /v1/config/sip-trunk/validate:
    post:
      operationId: validateFalePacoRegistration
      tags: [Fale Paco telephony]
      summary: Validate Fale Paco registration without making a call
      description: Reports safe credential presence, canonical domain and username matches, TCP transport, registration requirement, and current Baresip registration state. Does not return password or SIP authentication digests and does not send an INVITE.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: No-call registration readiness, content: { application/json: { schema: { $ref: '#/components/schemas/FalePacoSIPValidation' }, example: { ok: true, password_configured: true, credential_present: true, domain_match: true, username_match: true, realm_match: true, transport: tcp, registration_required: true, registration_state: REGISTERED } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/RuntimeFailure' }
  /api/v1/agent/prompt:
    get:
      tags: [Owner prompt]
      operationId: getActiveAgentPrompt
      summary: Read the active immutable agent prompt
      description: Returns the current active prompt version from the existing PostgreSQL prompt repository. New Gemini sessions snapshot it exactly once; changing the active prompt does not change an already-started session.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Active prompt, content: { application/json: { schema: { $ref: '#/components/schemas/PromptVersion' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No active prompt is configured }
    put:
      tags: [Owner prompt]
      operationId: createAndActivateAgentPrompt
      summary: Create and activate a new prompt version
      description: Creates a new immutable version using the existing PromptService and PostgreSQL repository, then activates that version. History is retained; older versions are never overwritten. Only sessions started afterward use this snapshot.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        description: Prompt identity and commercial instructions; do not put secrets or voice prosody settings here.
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromptDraft' }
            examples:
              Consultative SDR:
                value: { name: 'Joel SDR Voice V1', prompt: 'Conduza uma conversa consultiva em português brasileiro. Faça uma pergunta por vez. Não invente preço, agenda, resultados nem capacidades.' }
      responses:
        '201': { description: New active immutable version, content: { application/json: { schema: { $ref: '#/components/schemas/PromptVersion' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ProviderFailure' }
  /api/v1/agent/prompt/versions:
    get:
      tags: [Owner prompt]
      operationId: listAgentPromptVersions
      summary: List immutable prompt history
      description: Returns all stored versions newest first, including active state and timestamps. This is read-only and does not change running session snapshots.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Prompt versions, content: { application/json: { schema: { $ref: '#/components/schemas/PromptVersionList' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/agent/prompt/versions/{version}/activate:
    post:
      tags: [Owner prompt]
      operationId: activateAgentPromptVersion
      summary: Activate an existing prompt version
      description: Activates the requested immutable historical version for sessions created after this request. Already-running Gemini sessions keep their frozen snapshot.
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: version, in: path, required: true, description: Positive immutable version number from the versions endpoint., schema: { type: integer, minimum: 1 } }
      responses:
        '200': { description: Activated prompt, content: { application/json: { schema: { $ref: '#/components/schemas/PromptVersion' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Version does not exist }
  /v1/config/jev:
    get:
      tags: [Owner tuning]
      operationId: getJEVConfig
      summary: Read safe JEV tuning
      description: Returns process-local model, behavior description, owner guidance, configured_timeout_ms and canonical_default_timeout_ms (400). Values reset to defaults on API restart. OPENROUTER_API_KEY is read only from the local environment and is never returned.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Safe JEV settings, content: { application/json: { schema: { $ref: '#/components/schemas/JEVConfig' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [Owner tuning]
      operationId: updateJEVConfig
      summary: Update bounded JEV behavior tuning
      description: Updates the model and calibration text in process memory for future decisions; values reset to defaults on API restart. timeout_ms accepts 50–5000 ms; the canonical default remains 400 ms. Larger values can reduce false timeouts while increasing turn latency. Intent/action/reason enums, opt-out handling, tools, and hard safety rules are server controlled. JEV chooses what to do, not what to say. API keys are not accepted.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        description: All fields are required; unknown fields are rejected.
        content:
          application/json:
            schema: { $ref: '#/components/schemas/JEVConfigUpdate' }
            examples:
              Recommended:
                value: { model: 'typesafe/jev-1.13', timeout_ms: 400, description: 'Classificador comercial SDR responsável por selecionar a próxima ação.', decision_guidance: 'Classifique semanticamente o último turno FINAL e selecione a próxima ação comercial.' }
      responses:
        '200': { description: Applied JEV settings, content: { application/json: { schema: { $ref: '#/components/schemas/JEVConfig' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: Invalid model, field length, or timeout bound }
  /v1/config/gemini:
    get:
      tags: [Owner tuning]
      operationId: getGeminiConfig
      summary: Read safe Gemini Live voice settings
      description: Reads process-local model, prebuilt voice name, voice description, and delivery style. Values reset to defaults on API restart. The description and style are appended to a clearly delimited section of the frozen system instruction. No speech_metadata field is sent to Gemini Live.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Safe Gemini settings, content: { application/json: { schema: { $ref: '#/components/schemas/GeminiConfig' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [Owner tuning]
      operationId: updateGeminiConfig
      summary: Update Gemini Live voice settings
      description: Tunes how generated speech sounds for sessions/tests created after this update; settings live in process memory and reset to defaults on API restart. model selects Gemini Live; voice_name maps to generationConfig.speechConfig.voiceConfig.prebuiltVoiceConfig.voiceName; description and style are frozen into the system instruction. Do not put product facts, pricing, JEV rules, calendar data, tools, or secrets in style. Test-only overrides are not persisted.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GeminiConfig' }
            examples:
              Consultivo: { value: { model: 'gemini-3.8-live', voice_name: 'Kore', description: 'Voz comercial brasileira, humana e consultiva.', style: 'Calmo e confiante. Ritmo moderado. Frases curtas.' } }
              Executivo: { value: { model: 'gemini-3.8-live', voice_name: 'Kore', description: 'Voz profissional clara e segura.', style: 'Objetivo e seguro. Ritmo moderado, sem pressa. Preserve pausas naturais.' } }
              Acolhedor: { value: { model: 'gemini-3.8-live', voice_name: 'Kore', description: 'Voz cordial e atenta.', style: 'Próximo e cordial. Demonstre atenção ao lead sem soar ensaiado.' } }
              Calmo: { value: { model: 'gemini-3.8-live', voice_name: 'Kore', description: 'Voz serena e confiável.', style: 'Calmo, ritmo moderado, sem entusiasmo exagerado.' } }
              Objetivo: { value: { model: 'gemini-3.8-live', voice_name: 'Kore', description: 'Voz direta e profissional.', style: 'Frases curtas, dicção clara e uma pergunta por vez.' } }
              Moderadamente energético: { value: { model: 'gemini-3.8-live', voice_name: 'Kore', description: 'Voz brasileira positiva, sem tom publicitário.', style: 'Energia moderada e natural. Sem urgência ou entusiasmo exagerado.' } }
      responses:
        '200': { description: Applied safe settings, content: { application/json: { schema: { $ref: '#/components/schemas/GeminiConfig' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: Invalid model, voice name, or field length }
  /v1/test/jev:
    post:
      tags: [No-call diagnostics]
      operationId: testJEVDecision
      summary: Run one real JEV semantic classification without a call
      description: Sends one bounded lead_text to the configured real OpenRouter JEV. stage defaults to active and must be opening, active, closing, or ended. Returns only the safe typed intent, canonical action/reason, latency, timeout, and provider class. No raw provider payload, headers, credentials, or hidden reasoning are returned. JEV does not generate spoken copy.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiagnosticLeadRequest' }
            examples:
              Acceptance: { value: { lead_text: 'Sim, podemos marcar uma conversa para conhecer melhor.' } }
              Internal Alignment: { value: { lead_text: 'Preciso conversar com meu sócio primeiro.' } }
              Future Follow-up: { value: { lead_text: 'Me chama semana que vem.' } }
              Cost Objection: { value: { lead_text: 'O investimento ficou acima do que eu esperava.' } }
              Security Objection: { value: { lead_text: 'Tenho preocupação com LGPD e segurança.' } }
              Rejection: { value: { lead_text: 'Obrigado, mas não tenho interesse.' } }
              Opt-out: { value: { lead_text: 'Por favor, não entre mais em contato.' } }
              Human Request: { value: { lead_text: 'Quero falar com uma pessoa.' } }
              Clarification: { value: { lead_text: 'Como funciona esse serviço?' } }
              Capability Request: { value: { lead_text: 'Vocês integram com meu CRM?' } }
      responses:
        '200': { description: Safe semantic result; no side effects beyond one provider classification, content: { application/json: { schema: { $ref: '#/components/schemas/JEVTestResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { description: Unsupported conversation stage }
        '502': { $ref: '#/components/responses/ProviderFailure' }
        '504': { description: JEV exceeded configured timeout }
  /v1/test/agent-turn:
    post:
      tags: [No-call diagnostics]
      operationId: testAgentTurn
      summary: Run one real JEV-to-Gemini turn without telephony
      description: Performs one real JEV classification, builds the canonical TurnDirective, snapshots the active immutable prompt exactly once, then runs Gemini Live with the frozen voice configuration. Passes only when actual output transcription, PCM 24 kHz audio, generationComplete, and turnComplete are observed. Does not use SIP, Baresip, INVITE, or a call record. description/style request fields are temporary overrides and are not persisted. Raw PCM is retained in memory only for at most 15 minutes, bounded to 8 tests and 24 MiB total.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentTurnRequest' }
            examples:
              Privacy question: { value: { lead_text: 'Tenho preocupação com LGPD.' } }
      responses:
        '200': { description: Actual output transcription, lifecycle timings, and ephemeral audio link, content: { application/json: { schema: { $ref: '#/components/schemas/AgentTurnResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '502': { $ref: '#/components/responses/ProviderFailure' }
        '504': { description: Provider test exceeded its time limit }
  /v1/test/agent-turn/{test_id}/audio:
    get:
      tags: [No-call diagnostics]
      operationId: getAgentTurnAudio
      summary: Read an ephemeral diagnostic WAV
      description: Returns the actual Gemini PCM as mono signed 16-bit little-endian WAV at 24 kHz. This is generated diagnostic audio, not a call recording. The in-memory artifact expires within 15 minutes and is not persisted to PostgreSQL.
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: test_id, in: path, required: true, description: Test identifier returned from agent-turn., schema: { type: string } }
      responses:
        '200': { description: Mono PCM 24 kHz WAV, content: { audio/wav: { schema: { type: string, format: binary } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: Missing or expired ephemeral audio }
  /v1/runtime/status:
    get:
      tags: [Owner diagnostics]
      operationId: getRuntimeReadiness
      summary: Read safe local runtime readiness
      description: Reads Baresip ctrl_tcp registration state, active prompt identity, and configured provider model/voice readiness. Returns no credentials, SIP account contents, or provider payloads. WhatsApp and scheduling are explicitly deferred and non-blocking.
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Safe readiness metadata, content: { application/json: { schema: { $ref: '#/components/schemas/RuntimeStatus' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { description: Runtime readiness could not be established }
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Missing or invalid owner bearer token, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    RuntimeFailure: { description: Safe telephony runtime failure, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: string, description: Safe summary; raw provider messages are omitted. }, provider_status_class: { type: string, description: Optional sanitized provider operation/transport category. } } }
...
    FalePacoSIPPassword:
      type: object
      additionalProperties: false
      required: [secret]
      properties:
        secret: { type: string, title: Password, minLength: 1, writeOnly: true, description: Owner-supplied Fale Paco SIP password. Stored only in the protected local Baresip runtime profile. Baresip account syntax currently accepts up to 512 UTF-8 bytes and rejects semicolons, quotation marks, backslashes, and control characters to keep account parameters unambiguous. }
      example: { secret: YOUR_SIP_PASSWORD }
    FalePacoSIPProfile:
      type: object
      additionalProperties: false
      required: [provider, host, port, transport, registrar, outbound_proxy, auth, from_user, from_domain, caller_id, registration_required, enabled, password_configured]
      description: Canonical server-controlled Fale Paco profile. The only owner-supplied value is its write-only password; no codecs are advertised here because Baresip uses its configured runtime defaults.
      properties:
        provider: { type: string, const: 'Fale Paco', description: Provider is fixed by the server. }
        host: { type: string, const: '98034.falepaco.com.br' }
        port: { type: integer, const: 5060 }
        transport: { type: string, const: tcp }
        registrar: { type: string, const: '98034.falepaco.com.br' }
        outbound_proxy: { type: string, const: '98034.falepaco.com.br:5060' }
        auth: { type: object, required: [type, username, realm], properties: { type: { type: string, const: userpass }, username: { type: string, const: '100' }, realm: { type: string, const: '98034.falepaco.com.br' } } }
        from_user: { type: string, const: '100' }
        from_domain: { type: string, const: '98034.falepaco.com.br' }
        caller_id: { type: string, const: '551155200455' }
        registration_required: { type: boolean, const: true }
        enabled: { type: boolean, const: true }
        password_configured: { type: boolean, description: Whether a valid canonical account credential is present in protected local runtime storage. }
...
    PromptDraft:
      type: object
      additionalProperties: false
      required: [name, prompt]
      properties:
        name: { type: string, maxLength: 120, description: Human-readable prompt version name, example: 'Joel SDR Voice V1' }
        prompt: { type: string, maxLength: 32768, description: Immutable commercial persona, rules, and guardrails for Gemini sessions. }
...